Reorganize project structure
- Added `__init__.py` and `__main__.py`
- Package installer simplified and no longer part of main app
- `settings.py` is now alongside source code

# Conflicts:
#	dqt/settings_manager.py
#	dqt/tracker.py
#	main.py
