+++ dqt/settings.py
