+++ dqt/settings_manager.py
