+++ dqt/styletext.py
